+++ src/test/java/com/musseukpeople/woorimap/auth/infrastructure/JwtTokenProviderTest.java
@@ -1,4 +1,4 @@
-package com.musseukpeople.woorimap.auth.application;
+package com.musseukpeople.woorimap.auth.infrastructure;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
